# refactor(BinaryFile): swap and write binary data in buffered chunks