@Luthaf 2022-04-15
About the benchmarks: you have very similar values for cstdio and mmap here, I guess this is because of the tmpfs living in RAM. If you have some time to re-run some of it with a normal filesystem, there should be much more difference between the two methods (mmap should be similar to a RAM fs, but cstdio should be much slower).