@ezavod 2022-04-23
> If you have some time to re-run some of it with a normal filesystem, there should be much more difference between the two methods (mmap should be similar to a RAM fs, but cstdio should be much slower).

I did some automated benchmarking by running the initial test on different disks.
In general the changes in this PR make R+W faster than the current master. (Except on the ext_hdd, which I don't fully understand.)
Interestingly, now `cstdio` is always **faster** than `mmap`. And, for reading, the internal SSHD (i.e. HDD with SSD cache) is somewhat faster than the SSD and tmpfs. There might be some OS-level disk caching going on and fluctuations can be quite large.

<details>
<summary>Relevant HW info</summary>
Note int_hdd is a SSHD.

```
Disk ssd: 931,51 GiB, 1000204886016 bytes, 1953525168 sectors (BTRFS)
Disk model: WDC WDS100T2B0C-00PXH0                  
Units: sectors of 1 * 512 = 512 bytes
Sector size (logical/physical): 512 bytes / 512 bytes
I/O size (minimum/optimal): 512 bytes / 512 bytes
Disklabel type: dos


Disk int_hdd: 1,82 TiB, 2000398934016 bytes, 3907029168 sectors (BTRFS)
Disk model: ST2000LX001-1RG1
Units: sectors of 1 * 512 = 512 bytes
Sector size (logical/physical): 512 bytes / 4096 bytes
I/O size (minimum/optimal): 4096 bytes / 4096 bytes
Disklabel type: gpt


Disk ext_hdd: 3,64 TiB, 4000787025920 bytes, 976754645 sectors (external USB 3.0, NTFS)
Disk model: Expansion Desk  
Units: sectors of 1 * 4096 = 4096 bytes
Sector size (logical/physical): 4096 bytes / 4096 bytes
I/O size (minimum/optimal): 4096 bytes / 4096 bytes
Disklabel type: dos
```
</details>

![bench_perf-bin](https://user-images.githubusercontent.com/16089150/164909924-4dffe385-b282-45a5-9d76-197689ed3e65.png)