@Luthaf 2022-04-23
>  In general the changes in this PR make R+W faster than the current master. (Except on the ext_hdd, which I don't fully understand.)
>
> Interestingly, now cstdio is always faster than mmap. And, for reading, the internal SSHD (i.e. HDD with SSD cache) is somewhat faster than the SSD and tmpfs. There might be some OS-level disk caching going on and fluctuations can be quite large.


Both are excellent news! If we can do without the additional complexity of mmap, it would be great. I'll run some additional benchmarks on the machines I have access to, since the exact OS/FS used would also make a difference.