@Luthaf 2022-04-25
So one more datapoint for benchmarks, running the same code as in the netcdf PR:

```
Reading:
    mmap:
        master: 0.233 ms
        PR: 0.236 ms
    cstdio:
        master: 0.540 ms
        PR: 0.569 ms        

Writing:
    mmap:
        master: 0.713 ms
        PR: 0.330 ms
    cstdio:
        master: 2.76 ms
        PR: 0.744 ms
```

So for me read performance is unaffected, and write performance increases (dramatically so for cstdio). In both cases, mmap is still faster, so I would tend to keep it as a default for now.