@ezavod 2022-04-26
It's a little bit puzzling that writing for you is the the exact opposite as for me. At first I thought it's because `water.nc` is relatively small (~300 atoms) when compared to my benchmarks (50k atoms) and 100 frames vs 1k. But for such small files I see even up to 2-5x faster writes with `cstdio` when compared to `mmap`.
Anyway, as this seems to be heavily hardware/filesystem/OS dependent, so I agree with keeping things as they are for now.

<details>
<summary>For clarity, this is my benchmark setup</summary>
One can change the line `+#define CHEMFILES_BINARY_FILE_USE_MMAP 1` in the patch to switch between mmap and cstdio. And `NSTEPS` and `NATOMS` can be changed for in `main_read.cpp`.

```bash
#!/bin/bash

set -e

echo "checkout chemfiles ..."

if [ ! -d "chemfiles" ]; then
    git clone https://github.com/chemfiles/chemfiles.git
fi

cd chemfiles
git reset --hard
git checkout 17a14fca75cfdd8754c8d00b52674bbadec83029

echo "apply patch ..."

cat << 'EOF' > use_cstdio.patch
diff --git a/include/chemfiles/files/BinaryFile.hpp b/include/chemfiles/files/BinaryFile.hpp
index aa7d7942..d4d037a4 100644
--- a/include/chemfiles/files/BinaryFile.hpp
+++ b/include/chemfiles/files/BinaryFile.hpp
@@ -11,17 +11,7 @@
 
 static_assert(sizeof(char) == sizeof(int8_t), "char must be 8-bits");
 
-#if defined(__unix__) || (defined(__APPLE__) && defined(__MACH__))
-#if (CHEMFILES_SIZEOF_VOID_P == 8)
-    // use mmap in 64-bit posix
-    #define CHEMFILES_BINARY_FILE_USE_MMAP 1
-#else
-    // otherwise fallback to an implementation using only <cstdio>
-    #define CHEMFILES_BINARY_FILE_USE_MMAP 0
-#endif
-#else
-    #define CHEMFILES_BINARY_FILE_USE_MMAP 0
-#endif
+#define CHEMFILES_BINARY_FILE_USE_MMAP 1
 
 namespace chemfiles {
 
EOF

git apply use_cstdio.patch

echo "build chemfiles ..."
rm -rf build || true
mkdir build && cd build

cmake -DCMAKE_BUILD_TYPE=Release -DBUILD_SHARED_LIBS=ON -DCHFL_BUILD_TESTS=OFF CHFL_BUILD_DOCTESTS=OFF -DCHFL_TESTS_USE_VALGRIND=OFF ..
cmake --build . --config Release -- -j

cd ../..

echo "create benchmark source files ..."

cat << 'EOF' >> main_write.cpp
#include <iostream>
#include "chemfiles.hpp"

#define NSTEPS 1000
#define NATOMS 50000

int main() {
    chemfiles::Trajectory trajectory("out.nc", 'w');

    for (size_t step=0; step<NSTEPS; ++step) {
        auto frame = chemfiles::Frame(chemfiles::UnitCell({10, 10, 10}));
        for (size_t i=0; i<NATOMS; ++i) {
            frame.add_atom(chemfiles::Atom("H"), {static_cast<double>(i), -static_cast<double>(i%10), static_cast<double>((i*step)%10)});
        }
        trajectory.write(frame);
    }
}
EOF

cat << 'EOF' >> main_read.cpp
#include <iostream>
#include "chemfiles.hpp"

int main() {
    chemfiles::Trajectory trajectory("out.nc");

    std::cout << "There are " << trajectory.nsteps() << " steps in the trajectory" << std::endl;

    for (size_t step=0; step<trajectory.nsteps(); ++step) {
        auto frame = trajectory.read_step(step);
        (void)frame;
    }
}
EOF

echo "compile benchmarks ..."
chfl_path=chemfiles
export LD_LIBRARY_PATH=${chfl_path}/build

g++ -I${chfl_path}/include -I${chfl_path}/build/include -L${chfl_path}/build -lchemfiles main_write.cpp -o bench_write
g++ -I${chfl_path}/include -I${chfl_path}/build/include -L${chfl_path}/build -lchemfiles main_read.cpp -o bench_read


echo "run benchmarks ..."
hyperfine --warmup 3 ./bench_write
hyperfine --warmup 3 ./bench_read

echo "clean up ..."
rm -rf chemfiles
rm bench_read bench_write main_read.cpp main_write.cpp out.nc
```
</details>